Fix cropped logo in attribution control
The right side of the logo shown at the bottom left of the map was cropped.

First issue: The resolution of src/css/svg/maplibregl-ctrl-logo.svg was 91x23 but
the CSS expected 88x23. The logo has now been resized to the expected size.

Second issue: When the canvas container width is less than 250px, a
".mapboxgl-compact" style was supposed to crop the logo to a square 23x23 to keep
only its icon. This has been changed to 14x23 in the CSS to reflect the smaller
size of the icon in the MapLibre logo.