Update example links (#131)
This pull requests updates the links to examples in JSDoc comments from Mapbox to MapLibre. The links are defined in the ```@see``` tags and they show up in the [documentation website](https://maplibre.org/maplibre-gl-js-docs/api/).

I did not migrate all existing examples to the new documentation website because some of them were incompatible with ```maplibregl``` or they required a mapbox access token, see https://github.com/maplibre/maplibre-gl-js-docs/pull/36. Links to examples that were not migrated are removed.