Improve tile URL placeholder support (#348)
Fixes #346 

* Improve tile URL placeholder support

Currently tile URL placeholders are replaced only once in a URL. If the placeholder appears more than once, it additional placeholder doesn't get replaced. This is due to the replace function only replacing the first instance of a string. The simple solution is to us a global regex with the replace function. Performance wise there is little difference (regex is actually a tad bit faster in nearly all browsers).

I came across this issue on a project a couple years ago, but thought it was an edge case and workaround it with a transform request, but just saw this issue:  https://github.com/maplibre/maplibre-gl-js/issues/346 and see I'm not the only one who ran into this. This fix works for both raster and vector tile URLs.

* Update tile_id.test.js

Added test for tile_id improvement

* Update CHANGELOG.md

* Update CHANGELOG.md

* Update tile_id.test.js