moved elevation-calculation to GPU
This checkin is only for backup, so do not test!

* changed elevation calculation from CPU to GPU

This change had a big impact in performacne. To calculate the hight in
CPU sometimes about 100ms per tile was needed. When loading a lot of
tiles the framerate was really bad. This works now much better.
NOTE: Currently only the Mesh elevation is moved to GPU, symbols are
still calculated via the CPU.

* refactor render to texture workflow below ZL 14

Until now, below ZL 14 (e.g. our Vector-Tile maxzoom) the tile-size
increased with ZL, so ZL 1-14 had 1024px, ZL 15 2048px, ZL 16 4096px,
and so on. This was very easy to program, but needed also a lot of GPU
performacne and RAM. So now, the TerrainSourceCache holds tiles for each
zoomlevel of the size 1024, and below ZL 14 in each is renderd a
sub-region of ZL 14.

* refactor the coords-framebuffer.

Also the coords-framebuffer use a textures with 1024px in size. To avoid
bluring the texture below ZL 14 an additional small texture is needed
(u_coords_index) which holds in its RGBA values the tile quadrants of
the sub-regions.
NOTE: This is currently a work in progress!