account for one less control when LogoControl not shown by default
setting maplibreLogo: false in mapoptions (map.ts) means one less control exists on the default map. I have updated the map tests to account for one less control. I have also set the default to false in the util.ts so when we change this later it won't affect the tests.