correct coveringTiles when terrain is enabled (#1300)
* Fixes #1241 - correct coveringTiles when terrain is enabled

* fix lint

* Add unit test for calculating min/max elevation (#1316)

* Add unit test for calculating min/max elevation

* Use correct method name

* Move min/max elevation value calc to Terrain class

* Alter Terrain.getMinMaxElevation to take tileID arg

* Fix unit tests after merge from main

* Fix lint

* Add render test for 3D terrain (#1320)

* Add terrain style spec integration test

Tests that the root level terrain object must be in the following form:

    "terrain": {
      "source": string,
      "exaggeration": number,
      "elevationOffset": number
    }

* Add work-in-progress render test for 3D terrain

* Fix terrain render test so it passes

The original test was failing because there were no terrain tiles
available at the zoom being requested. Now it's testing at zoom 13,
which means requesting terrain tiles at zoom 12, which are available to
the tests.

* Update terrain render test to test fix for #1241

The fix for #1241, commit 352bc03e, ensures that elevationOffset is
taken into account when gauging which tiles are required to render 3D
terrain. If we add an extreme offset then we have a test that fails on
the current main branch but passes after the fix.

* Add changelog comment

Co-authored-by: Matt Riggott <flother@users.noreply.github.com>
Co-authored-by: HarelM <harel.mazor@gmail.com>