Change how meta key is detected for coop. gestures
UIEvents like 'wheel' include properties for whether some keys are
currently pressed, including ctrl and meta. [0] This should be less
prone to error, specifically when the user presses or depresses one of
these keys while the browser document is not active.

Adds a debug page for cooperative gestures because this feature cannot
be tested with a fullscreen map, and the only other test page is
a fullscreen map.

[0] https://w3c.github.io/uievents/#dom-mouseevent-ctrlkey