Add method to enable/disable cooperative gestures (#2152)
* Change how meta key is detected for coop. gestures

UIEvents like 'wheel' include properties for whether some keys are
currently pressed, including ctrl and meta. [0] This should be less
prone to error, specifically when the user presses or depresses one of
these keys while the browser document is not active.

Adds a debug page for cooperative gestures because this feature cannot
be tested with a fullscreen map, and the only other test page is
a fullscreen map.

[0] https://w3c.github.io/uievents/#dom-mouseevent-ctrlkey

* Add method to enable/disable cooperativeGestures

Fixes #2057

Also disable cooperative gestures in fullscreen, using these new methods
(Fixes #1488)

* Changelog

* Combine imports from map

The syntax is a bit odd. Background here:
https://github.com/tc39/proposal-type-annotations/issues/16

* Remove debug page

* Specify event type

* Add Map._getMetaKey()

* Add cooperative gestures class

* Add return value to _getMetaKey()

* Add _metaKey property, initialize in constructor

* Update expectedBytes to size of bundle on main