Add check method for terrain collision
For further render/browser test, we need a method to get information
about collision between camera and terrain.

Map.checkTerrainCollision() only returns object contains altitude of
camera and terrain surface with buffer. If the first one
(cameraAltitude) is lower than the second one (minAllowedAltitude), then
collision happens.

The third value in object is maxPitch, which means the maxPitch to keep
camera above minAllowedAltitude (if camera stays at the same position).
It is used to change Transform._pitch to reset camera.

The reason why not directly change Transform._elevation is that: it also
changes the map center if Transform._pitch remains. So user may be
surprised when they changing the bearing based on the map center.

In this case, changing Transform._pitch promise map center won't quickly
changes in the next few frames.