Introduce max canvas size option and automatically limit pixel ratio (#2674)
* rendering test: implement cropping of png result

This is useful for renderer results that are too large to be handled by puppeteer

* mock webgl: update drawing buffer size when viewport changes

* Round down painter and canvas dimensions so they're consistent

* add veryhigh pixel ratio test

This test draws a black square,
The pixel ratio and width were chosen to exceed the maximum texture value, which is set to 8192 in the CI environment.
The result is cropped due to the limits of the intercommunication between chromium and puppeteer.

* introduce maxCanvasSize option

A canvas size over GL MAX_TEXTURE_SIZE can cause distortions and an excessive canvas size can cause hangs or crashes.
Limiting canvas size can fix these issues. Add an option to let the users alter this limit.
The default value is [4096, 4096] as MAX_TEXTURE_SIZE is usally at least 4096px.
In CI max_texture_size is 8192, so we set maxCanvasSize to that value in tests.

* Resize map after setting pixel ratio

This has two benefits:
1. The map is immediately re-rendered with the new pixel ratio
2. We avoid duplicating pixel ratio clamping logic

* further clamp pixel ratio after hitting gl limits

Althoug setting maxCanvasSize to a value <= GL.MAX_TEXTURE_SIZE solves most problems, there is no guarantee that we cannot exceed other GL limits.
To handle these cases, we check that the drawing buffer is of the requested size, if it isn't we scale down pixel ratio further

* map test: move webgl error in its own section

* add a unit test for webgl drawing buffer limits

* update maxCanvasSize type and change comment type

* update maxCanvasSize after hitting gl limits

This is an optimization to avoid failing the overlimit check multiple times.

* update changelog