Tests for fix for #2094 (unwanted zoom changes)
Added asserts that camera altitude remains invariant throughout the
recalculateZoom test. These new asserts fail before the fix for 2094.

Moved the asserts that the zoom is a particular value to after the
altitude asserts - a particular zoom is correct if and only if it
results in the correct altitude, so whether the altitude assert fails
is more important information to have in a failing test.

Updated CHANGELOG.md