Refactor cooperativeGesture usage to colocate all logic
Previously, the logic for preventing gestures and notifying the
user about those gestures was duplicated. Firstly, we checked if
the gesture should go through by things like counting touches in
other handlers. We would also check the same thing in the CoopGesture
handler class, too. This was a source of confusion for me as it
wasn't apparent how the Handler was able to prevent other handlers
from firing - the answer is that it wasn't and the the other
handlers had to check with the CooperativeGestureHandler themselves.

Now, this is made explicit as the logic for checking only exists
in one place, and the other handlers are responsible for asking
the CooperativeGestureHandler to notify the user about the fact
that the gesture was blocked.