Fix multipolygon and point intersections
multiPolygonContainsPoint was sharing variables between
individual polygons inside the multipolygon. This left
a situation where an even number of polygons that
overlapped a single point, would be said to not contain
that point.

Now a multipolygon is said to contain a point if any
polygon contained in the multipolygon contains that
point.